+++ plugins/Admin/Admin.php
@@ -749,7 +749,7 @@ class Admin extends Plugin implements SplObserver, GetContentStrategyInterface, 
       "cacheable" => true,
     ];
     $vars["changestatus"] = [
-      "value" => $this->dataFileStatus == self::STATUS_DISABLED ? _("enable") : _("disable"),
+      "value" => $this->dataFileStatus == self::STATUS_DISABLED ? _("Enable file") : _("Disable file"),
       "cacheable" => true,
     ];
     $vars["changestatusurl"] = [
@@ -772,7 +772,7 @@ class Admin extends Plugin implements SplObserver, GetContentStrategyInterface, 
         "cacheable" => true,
       ];
       $vars["nohide"] = [
-        "value" => "nohide",
+        "value" => "hideable-nohide",
         "cacheable" => true,
       ];
     }