+++ plugins/Agregator/DocList.php
@@ -47,7 +47,7 @@ class DocList extends AgregatorList {
     $cacheUpTodate = false;
     $cache = null;
     $listDoc = null;
-    if ($cacheExists) {
+    if (Cms::getLoggedUser() != SERVER_USER && $cacheExists) {
       $cache = apc_fetch($cacheKey);
       $cacheUpTodate = $cache["newestFileMtime"] == $newestFileMtime;
     }