+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.12.0] - 2018-05-16
+### Added
+ - GlobalMenu has own transformation and styles
+ - Honeypot label
+ - Inputvar supports required attribute
+
+### Changed
+ - Align input description to input
+ - InputVar form inputs has no max-length
+ - Eventable supports any element
+ - Refactor hideable to use data attributes
+ - Improve scrolltopable UX
+
 ## [2.11.5] - 2018-06-10
 ### Removed
  - Default agregator.xsl
@@ -380,6 +393,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Plugin LinkList creates and displays local variable ``linklist``.
 - System messages from previous request are now marked via class.
 
+[2.12.0]: https://bitbucket.org/igwr/cms/compare/v2.12.0..v2.11.5
 [2.11.5]: https://bitbucket.org/igwr/cms/compare/v2.11.5..v2.11.4
 [2.11.4]: https://bitbucket.org/igwr/cms/compare/v2.11.4..v2.11.3
 [2.11.3]: https://bitbucket.org/igwr/cms/compare/v2.11.3..v2.11.2